clean non-annotate tags commit message
most of the info in the message are incorrect and it is complex
to gather them because the tag is written on disk only after the
update hook is executed. In most cases it was not showing
the correct diff either.

Drop all references to commit and spin a simple tag message instead.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>